Support for parallelizing QuadTopology and DualTopology energy evaluations; for ttOSRW, testLambdaGradient, and new multiTopTimer scripts, specify -np 1, 2, or 4 to split 1, 2, or 4 ways; -np 2 with quad-topology will instruct the QuadTopology to parallelize its two DualTopology energies, while -np 4 also instructs the component DualTopology objects to parallelize their ForceFieldEnergy energies. Prior parallelization is mostly unaffected, although number of SMP threads assigned to each ForceFieldEnergy is divided by -np.
To get this to function properly, a code path has been instituted to open files with a non-default number of threads; this can occur via PotentialsFunctions.open(file[s], nThreads) call, or via MainPanel.openWait(file[s], nThreads), as well as the Groovy method closure open(file[s], nThreads). This is implemented for all FFX implementations of PotentialsFunctions, and there is a default defender method in the interface which calls open without nThreads.

The multiTopTimer script has been added to permit timing multi-topology energies. To get this to function, two methods have been added to the Potential interface, energy(x[], boolean verbose) and energyAndGradient(x[], g[], boolean verbose); this is implemented for ForceFieldEnergy, DualTopologyEnergy, QuadTopologyEnergy, and the interface has a default implementation which simply calls energy or energyAndGradient without the verbose flag.