Added support for improved integration of dUdL for Thermodynamics, using the ttosrw-integrationType flag, which defaults to using Simpson's rule for quadratic interpolation.
Value of dUdL at 0 and 1 for these rules is either based off linear interpolation from the first and last 2 bin midpoints, or in some cases, exactly 0.

The dEdLZeroAtEnds method has been added to LambdaInterface, returning true if this LambdaInterface's dUdL is guaranteed to be 0 at its endpoints, such as for DFF with a trigonometric/multiplicative switch. If this returns true, the numeric integration scheme above can safely assume the values of dUdL at 0 and 1 are exactly 0. The default implementation returns false.