Probable bugfix to rotamer optimization (the ManyBody script). Previously, the currentEnergy method had the option to catch force field exceptions and return 1E100. This led to unintended behavior when it came to calcuating pairs and triples; if pruning was off, and a pair of rotamers both had unreasonable conformations that caused exceptions in energy calculation, the stored pair energy would be (1E100 - 1E100 - 1E100 = -1E100).
This option to catch exceptions has been removed, and every call to currentEnergy must have its own logic to catch ArithmeticExceptions propagating from the energy calculations. In the case of self/pair/triple/quad calculations, the energy term to be stored in the energy matrices is set to 1E100, which was the original intent of catching the ArithmeticExceptions (so that term would be quickly eliminated in pruning/DEE).