Frozen neighbor lists and GK cutoffs should be working now.
GK_CUTOFF is now a new variable, though it defaults to EWALD_CUTOFF.

With a regular, non-frozen neighbor list, the replicate cell is based on the maximum of the vdW, PME, and GK cutoffs. Specifying NEIGHBOR_LIST_CUTOFF in this case is an error.

With a frozen neighbor list, the replicate cell is based on the maximum of the vdW, PME, and neighbor-list cutoffs, with GK_CUTOFF allowed to exceed the neighbor-list cutoff. In this case, GK is calculated on atoms within GK_CUTOFF Angstroms, but only among that subset that was within NEIGHBOR_LIST_CUTOFF Angstroms at the start of the simulation.

In this case, specifying a neighbor-list cutoff < max(PME, vdW) is an error.

vdW cutoff continues to be >= PME cutoff.