Addition of a test class for NewThermodynamics; currently only tests the help message. To support this, the NewThermodynamics.groovy script can have an additional Configuration passed to it before run is called, and OSRWOptions.constructOSRW now telescopes to permit addition of a Configuration (or null).
The structure adopted by ThermodynamicsTest may be useful for parametrized tests of scripts with huge numbers of options. Static final unmodifiable maps of default options, properties, and flags are constructed. Each parametrized test starts with copies of these maps, then can over-ride these on an option-by-option basis.

This permits a test to change any option or any property individually, without needing to specify the ~45 options in the NewThermodynamics script for each and every parametrized test.