Created an AtomSelectionOptions Mixin for FFX Groovy Scripts, which includes support for selecting active and inactive atoms. Similar to Tinker, the inactive selection is evaluated first. Then, if any atoms are selected to be active, the latter takes precendence.
The syntax of the active / inactive options are [NONE, ALL, Range(s): 1-3,6-N], where the N is replaced by the number of atoms in the system. The default is that the inactive and active fields are both empty (i.e. all atoms are active if neither options is used).

Coupling active / inactive command-line options to dual-topology simulations has not been completed.

The active/inactive syntax is now also used to specify alchemical atoms and uncharged atoms for single and dual-topology simulations. The default is that no atoms are alchemical, and no atoms are uncharged.

Added a GeneralizedKirkwoodTensor class where methods to automatically generate Generalized Kirkwood interaction tensors will be placed.